chore(compose): the dedupe evidence decoding takes its own file under the length cap
